backend/requirements.txt
fastapi
sse-starlette
aiofiles
jinja2
uvicorn[standard]
requests_async
pyyaml
